method "computeRadonTrafoAndDerive()" moved
moved from SimpleOperators to EpipolarConsistency
